backend/requirements.txt
alembic==1.15.2
aniso8601==10.0.1
annotated-types==0.7.0
anthropic==0.25.9
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
coverage==7.8.0
distro==1.9.0
filelock==3.18.0
flask==2.3.2
flask-cors==3.0.10
flask-migrate==4.0.4
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
fsspec==2025.3.2
greenlet==3.2.2
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
mako==1.3.10
markupsafe==3.0.2
mock==5.1.0
openai==1.30.1
packaging==25.0
pip==25.1.1
pluggy==1.6.0
prometheus-client==0.20.0
psycopg2-binary==2.9.6
pydantic==2.7.1
pydantic-core==2.18.2
pytest==8.2.0
pytest-cov==5.0.0
python-dotenv==1.0.1
pytz==2025.2
pyyaml==6.0.2
requests==2.31.0
responses==0.25.7
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.16
tokenizers==0.21.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
werkzeug==3.1.3
wheel==0.45.1
